+++ cup_test/cup_import_test.py
@@ -13,8 +13,8 @@ _NOW_PATH = os.path.dirname(os.path.abspath(__file__)) + '/'
 sys.path.insert(0, _NOW_PATH + '../')
 
 
-from cup.thirdp import pymysql
-pymysql.install_as_MySQLdb()
-import MySQLdb
+# from cup.thirdp import pymysql
+# pymysql.install_as_MySQLdb()
+# import MySQLdb
 
 # vi:set tw=0 ts=4 sw=4 nowrap fdm=indent