+++ cup_test/manual/cup_storage_object_test.py
@@ -6,6 +6,7 @@
 :description:
     unittest for cup.cache
 """
+from __future__ import print_function
 import os
 import sys
 
@@ -15,27 +16,27 @@ sys.path.insert(0, _NOW_PATH + '../')
 from cup import unittest
 from cup.storage import obj
 
-def test_s3_object():
-    """test s3 object"""
-    config = {
-        'endpoint': 'abcd',
-        'ak': 'xxxx',
-        'sk': 'yyyy',
-        'bucket': 'cup-test'
-    }
-    s3obj = obj.S3ObjectSystem(config)
-    print s3obj.delete_bucket('cup-test', forcely=True)
-    unittest.assert_eq(s3obj.create_bucket('cup-test')['returncode'], 0)
-    unittest.assert_eq(
-        s3obj.put('test-obj', './cup_storage_object_test.py')['returncode'],
-        0
-    )
-    unittest.assert_eq(
-        s3obj.head('test-obj')['returncode'],
-        0
-    )
-    s3obj.get('test-obj', './test-obj')
-    os.unlink('test-obj')
+# def test_s3_object():
+#     """test s3 object"""
+#     config = {
+#         'endpoint': 'abcd',
+#         'ak': 'xxxx',
+#         'sk': 'yyyy',
+#         'bucket': 'cup-test'
+#     }
+#     s3obj = obj.S3ObjectSystem(config)
+#     print(s3obj.delete_bucket('cup-test', forcely=True))
+#     unittest.assert_eq(s3obj.create_bucket('cup-test')['returncode'], 0)
+#     unittest.assert_eq(
+#         s3obj.put('test-obj', './cup_storage_object_test.py')['returncode'],
+#         0
+#     )
+#     unittest.assert_eq(
+#         s3obj.head('test-obj')['returncode'],
+#         0
+#     )
+#     s3obj.get('test-obj', './test-obj')
+#     os.unlink('test-obj')