+++ docs/gitbook_src/zh-hans/ChangeLog.md
@@ -1,8 +1,10 @@
 # 1. Versions Released
-## Version 3.2.20
+## Version
+## Version 3.2.20 ~ 
 * [Compatibility] *cup.log* | Fix unicode bug for py3
 * [Enhancement] *cup.log* | Add xdebug/xinfo/xinit_comlog in order to support different loggers
 * [Removal] *cup.jenkinslib* | Recommand using python-jenkins to replace cup.jenkinslib
+* [Enhancement] *cup.cache*| Enhance KVCache, add pop_expired. Support setting up "maxsize" for the cache pool
 
 ## Version 3.2.19
 * [Enhancement] *cup.services.generator* | Add visibility to Cached UUID